Appears as if ADC is working

+++ source/hal/adc.h
@@ -1,4 +1,4 @@
-#include <stdint.h>
+#include "arm_math.h"
 
 void init_adc();
-void read_adc(uint32_t* reading);
+void read_adc(float32_t *reading);
